REF refactor traits
